pyproject.toml
[tool.poetry]
name = "Api"
version = "0.1.0"
description = "Api for backend Ghostpost Application."
authors = ["Mike Boring <shekinah999@msn.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
djangorestframework = "^3.12.1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"
flake8 = "^3.8.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
